Made players face correct direction with movement. Did some refactoring of file names so the first player has a player1 designation rather than just player
