+++ Assets/Scripts/Player1Controller.cs
@@ -1,15 +1,15 @@
-using System;
+﻿using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Player2Controller : MonoBehaviour
+public class Player1Controller : MonoBehaviour
 {
 
     public float moveSpeed = 5f;
     public int tilesMoved = 0;
 
-    public bool p2Turn;
+    public bool p1Turn;
 
     public Transform movePoint;
     public int SpeedInt;
@@ -38,24 +38,24 @@ public class Player2Controller : MonoBehaviour
     {
         transform.position = Vector3.MoveTowards(transform.position, movePoint.position, moveSpeed * Time.deltaTime);
 
-        if (p2Turn == true)
+        if (p1Turn == true)
         {
             if (Vector3.Distance(transform.position, movePoint.position) <= 0.05f)
             {
-                // Set player direction
-                if (Input.GetAxisRaw("Horizontal") > 0 && !facingRight)
-                {
-                    Flip();
-                }
-                else if (Input.GetAxisRaw("Horizontal") < 0 && facingRight)
-                {
-                    Flip();
-                }
 
-                // Move player
                 if (Mathf.Abs(Input.GetAxisRaw("Horizontal")) == 1f)
                 {
+                    // Set player direction
+                    if (Input.GetAxisRaw("Horizontal") > 0 && !facingRight)
+                    {
+                        Flip();
+                    }
+                    else if (Input.GetAxisRaw("Horizontal") < 0 && facingRight)
+                    {
+                        Flip();
+                    }
 
+                    // Move Player
                     if (!Physics2D.OverlapCircle(movePoint.position + new Vector3(Input.GetAxisRaw("Horizontal"), 0f, 0f), 0.2f, whatStopsMovement))
                     {
                         movePoint.position += new Vector3(Input.GetAxisRaw("Horizontal"), 0f, 0f);
@@ -82,8 +82,8 @@ public class Player2Controller : MonoBehaviour
 
             if (tilesMoved == 5)
             {
-                //p2Turn = false;
-                //Debug.Log("Player 2's turn over!");
+                //p1Turn = false;
+                //Debug.Log("Player 1's turn over!");
                 //tilesMoved = 0;
             }
         }
