+++ Assets/ghostMove.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a44649ba098ffa946955150923128d9f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
